chore(refactor): migrate to ts + use dossier structure
