+++ package.json
@@ -7,12 +7,13 @@
   "private": true,
   "engines": {
     "yarn": ">=1.7.0",
-    "node": "=16.x"
+    "node": ">=16.x"
   },
   "scripts": {
     "lint": "eslint .",
     "test": "jest",
-    "start": "babel-node index.js"
+    "start": "ts-node src/index.ts",
+    "start:dev": "nodemon src/index.ts"
   },
   "devDependencies": {
     "@babel/cli": "^7.1.0",
@@ -27,6 +28,9 @@
     "eslint-plugin-prettier": "^2.6.2",
     "jest": "24.8.0",
     "prettier": "^1.14.3",
-    "regenerator-runtime": "^0.12.1"
+    "regenerator-runtime": "^0.12.1",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3",
+    "nodemon": "^3.0.2"
   }
 }
