+++ src/tests/pharmacy.test.js
