+++ tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "compilerOptions": {
+    "target": "es6",
+    "module": "commonjs",
+    "outDir": "dist",
+    "downlevelIteration": true,
+    "skipDefaultLibCheck": true,
+    "skipLibCheck": true,
+    "strict": true,
+    "typeRoots": ["./types", "./node_modules/@types"],
+    "esModuleInterop": true
+  },
+  "include": ["**/src/**/*.ts"],
+  "exclude": ["tests"]
+}
